Initial Commit
QOM (Quality Object Model) RPG sim and GM tracking engine
Author: Woody Phillips (PistolBear)
Files: src folder and all subfolders
Notes:  Initial check in, should have a good baseline for Abilities,
classes, and in particular use for me right now, the RollInit class,
which uses a simple Swing based interface to quickly set up a re-useable
system for rolling initiative for N players, in a way that re-rolls can
be done as long as it is not closed.

Next Steps:
1. Allow for custom (actual) user/player names.
2. Allow for export of saved player initiative information.
3. Define a more robust interface with things like Temporary mods should
be allowed for when players use initiative-modifying abilities,
notifications of "Holding Action", and others.